refactor (challenge): dto 계층 이동
